requirements.txt
requests==2.27.1
Flask==2.0.2
gunicorn[gevent]
gevent
azure-identity>=1.12.0
azure-keyvault-secrets>=4.7.0
